Hgfs Server: Fix a memory leak on server exit not destroying the shares
This change fixes a memory leak on server exit not destroying the shares.
The list of shares is not destroyed and left.

Fix this at the Hgfs server tear down by calling the shares reset with an
empty list which will cause all shares to be treated as stale and so deleted.