Create symlink for vmware-user
When the vmtoolsd user daemon receives a SIGUSR2, it restarts itself
by executing 'vmware-user' and terminates. This is used on upgrades,
to make sure the current version of vmtoolsd is running.

In the tar installer, /usr/bin/vmware-user is a symlink to
vmware-user-suid-wrapper. Unfortunately, that symlink was not installed
by open-vm-tools.

Package maintainers should send this signal to the vmtoolsd user
process when open-vm-tools is updated.