Hgfs OSX Server: Reinstate checkpoint synchronization and process session
First reinstate the server checkpoint synchronization routine by adding
it into the table the HGFS server session callback table which is returned
to the transports upon initialization. These transports will then call
into the HGFS server when asked by the HGFS server manager which receives
the check point synchronization calls.

When processing the synchronization request for the operation type -
freeze or thaw we must pass through the session information to the
activation and deactivation calls into the file system change notification
subsystem. These calls will require the session for register-unregister
any active threads generating events.  This change just plumbs through
the session as the argument. Follow on changes will make use of the
session argument passing it on to the event generating component.

Bonus here is that we can delete an unwarranted global from the HGFS server.
A to do comment is added to move the async globals into the HGFS server
session object as they should be a per HGFS  session and are useful only
on a per session basis. No session then no need for them to be around and
if the session does have asynchronous support then again will not be required.
For example, if the transport does not support this behavior no need to
check them.