Remove Util_GetCurrentThreadId from VIX
Some cleanup now that VThreadBase_GetKernelID() is available
in VIX code, we log the thread which made certain API calls.

The thread name is already logged in our Log() implementation,
so having the thread ID merely wastes bytes. Remove it.

(Thread IDs like this are useful for correlating with kernel logs.
Which in practice does not apply to VIX).