Tools Hgfs Plugin: Refresh Hgfs drives on client start
In order to get Explorer to see that VMware Shared Fodlers
drives are available when the feature is enabled, we touch
each drive that is mapped as a VMware Shared Folder. This
is to prevent users being confused when they have the feature
enabled but Explorer creates the user's drive mappings before
the tools service has initialized the Hgfs plugin and started
the Hgfs client. In this situation the drives are marked
as red (disconnected) until the user clicks on the mapped
drive when Explorer will refresh the connction and set it
to green (connected).

Details:
- Main service will usually create otheriwse open a named event
  to synchronize with the user service.
- Main service will start the HGFS client redirector (as before
  this is not new).
- Main service will set the synchronization event to signalled.
- User service will usually open the named event (or create it
  if first).
- User service will enumerate all the mapped network drives and,
  if they are provided by the VMware Shared Folder redirector,
  reconnect them.
- The named event will be created with Administrators group all
  access and Everyone with modify state to signal it and synchronize
  to wait on it.