Hgfs OSX Server: Create and use server callback table for change notifications
There are now three different callbacks into the Hgfs server from the file
change notification component. So we now pass these in a table to the nofify
module when it is initialized instead of ad hoc. Previously, the same
callback is passed for every subscriber watch event and therefore duplicated
and stored in each subscriber.

Note that the feature is currently switched off by default until these changes
are complete. A user will explicitly have to enable the VMX file setting and
would cause a VMX assert to be triggered.

Details:
- The server now creates a table for its notification callbacks and passes
  it to the notify initialization function.
- The server now includes the async thread registration callbacks previously
  ifdef'd out.
- The Notify_Init function APIs are updated to reflect the new argument for
  callbacks.
- The HgfsNotify_AddSubscriber argument does not include the callback as it
  is now passed once at the initialization time.
- The HgfsNotifyEventReceiveCb type has been modified now it is an entry in
  the callback table.
- The subscriber callback function call is now validated to be non-NULL
  before calling.
- The Mac notify module adds the server callback table to its FS event
  context which is passed to the event generating submodule for the Mac
  implementation.
- The Mac notify module adds its own FS event callbacks for thread register
  and unregister called from the FS event module when registering and
  unregistering its FS event thread.