lib/file: Continue simplifying and optimizing
Since we're UTF8 only, there are simplifications and
optimizations that can be done to lib/file, particularly in the
routines that look for directory separators and ASCII characters like
'.'. We don't need to use the unicode library to deal with these
type of code points.

This is the second in a sequence of changes to improve lib/file.