Tools Hgfs Plugin: refresh Hgfs drives on client start
In order to get Explorer to see that VMware Shared Folders drives are
available when the feature is enabled, we touch each drive that is
mapped as a VMware Shared Folder.  This is to prevent users being
confused when the have the feature enabled but Explorer creates the
user's drive mappings before the tools service has initialized the
Hgfs plugin and started the Hgfs client.  In this situation the drives
were marked as red (disconnected) until the user clicks on the mapped
drive.  Then Explorer will refresh the connction and set it to green
(connected).