Hgfs Protocol: clean up the Hgfs protocol operation capabilities
This change cleans up the operation capability and its flags which
currently are not clearly associated with the type to which they refer.
The name doesn't help in that regard either.

The capability and flags are used in the Hgfs session creation by
both the server and client to dictate the protocol operation versions
to be used during the lifetime of that session. The flags denote
which operations are currently used and how they are expected to be
used.