Fix ASSERT of registering the same callback twice
The previous change for Poll introduced a bug in AsyncSocket.
When re-registering the recv callback, it needs to check whether
there is a concurrent blocking recv (which means the send callback
has detected a disconnection and is draining the pipe).  The recv
callback should act like a no-op in that case (other than to
release the reference).