GuestInfo: enable nic/iproute information for Linux without libdnet
Linux glibc (since 2.3) fully supports getting ip addresses using
getifaddrs() for both IPv4 and IPv6. Use of getifaddrs() was already
implemented in nicInfoPosix.c for tools for ESXi. This change
enables that code also for Linux, if libdnet is not used.

The change also enables the iproute information, which does not
depend on libdnet, except by the use of addr_stob(), which can
be replaced easily with the existing function CountNetmaskBits().