Add SUSE Server for SAP Applications to hostinfoPosix.c
On SLES-SP1-SAP the "lsb_release -sd" command returns the string
"SUSE Linux Enterprise Server for SAP Applications 12 SP1".
Parsing in HostinfoGetOSShortName() detects the sub-strings "suse"
and "enterprise" but not "server 12".  STR_OS_SLES is returned
resulting in displaying of "SUSE Linux Enterprise 8/9" in the
summary screen of the vSphere client.  This is wrong.

Instead, search for the sub-string "server for sap applications 12"
and return STR_OS_SLES_12.  SLES11-SAP is not affected.

Please see: github.com/vmware/open-vm-tools/pull/123