GuestInfo: add primary-nics and low-priority-nics options
Add the options 'primary-nics' and 'low-priority-nics' for GuestInfo.

The options accept a list of comma separated patterns for interface names
which can be considered as 'primary' or 'low priority' network interfaces.
This will cause the NicInfo to be put on top of the list for primary
interfaces, and to the bottom for low priority interfaces.

Example:

primary-nics=eth1

will make sure that the IP address(es) for eth1 will be sorted on
top of the list of IP addresses.

primary-nics=eth*

will make sure that any one of the addresses of all interfaces
matching eth* will be sorted on top of the list of IP addresses,
but won't guarantee which one if there is more than one.

low-priority-nics=eth*

will work analogously, but interfaces will be sorted to the bottom.

In case the limit of the number of interfaces to be reported is reached,
low priority interfaces are the first to be skipped.