Tools: GuestInfo: ESX: dynamically load libvmkmemstats.so
The new version of the vmware tools will be installed on older version
of ESX.  But these older versions do not come with libvmkmemstats.so
and a dynamic linking error occurs.

This patch changes the code to deal with the case where the lib is not
available.  Instead of relying on the linker to dynamically link with
libvmkmemstats.so, g_module_open() is used to open the library, if
available, and gracefully handle any eventual errors.
Before the patch, the whole libguestinfo would stop working and after
the patch only the memory stats are unavailable.