Allow copying of object policy within the same container
ObjLib_CloneExtParams is currently not allowing copying of policy from
parent to child/clone object on VSAN. This is because the copy policy
check is based on unique filesystem ID, which will be different for
parent and child on VSAN because they will mostly be created on same
container but different namespace. This change uses VCFS root instead
of namespace UUID to make the comparison on VSAN.

Bonus fix: File_FullPath is now fixed to work for any combination of
ancestor path being present. Currently it only works if either the
given path exists, or when only the last component of the path does
not exist. And since this fix relies on finding the VCFS root of both
parent and child (parent is present but child is not),
ObjLibPathToVCFSRoot has been modified to rely on File_FullPath  in
addition to Util_GetCanonicalPath if the path does not exist.