Remove some __STRICT_ANSI__ guards in vm_basic_types.h
Some time in the distant past, these guards were needed to
work around... something. (Probably, it was that 'long long'
is not a valid type in strict c90 mode). Note that we don't
use c90 mode, we only use gnu90 or c99.

This ends up being relevant because for C++ we are indeed
using -std=c++11 (which implies __STRICT_ANSI__). The Linux/C++
section of this file thus had duplicate definitions (which I am
now removing) to work around the gap.

Quick cheat sheet for -std= values:
- c90 := __STRICT_ANSI__, no long long type
- gnu90 := c90 + gcc extensions, has long long type
- c99 := __STRICT_ANSI__, has long long type (defined in c99)
- c++11 := __STRICT_ANSI__, has long long type (defined in C++11)