vm_basic_types.h: use stdint.h a little more.
Now that enough of the builds are C++11, let's make the common case
use the <stdint.h> header to define the intNN types directly
from language-standard intNN_t types, and treat pre-C99/C++98
code as the special case.

This also provides some clarity about the last __STRICT_ANSI__
macros in this file.  One was related to usage of 'long long' in
strict c90 mode and has been replaced with a comment indicating that
64-bit-variable compiler extensions are mandatory (strict c90 mode is
not allowed via the headers).  The other related to X11 headers; those
headers have been examined to understand the workarounds needed and
the exact scenario(s) that are problematic have been commented (turns
out FreeBSD had nothing to do with the problem).

Finally, all the Bool-related definitions have been moved and merged
to a single location NOT at the top of the file.

Net effects of this change:
- non-linux C++11 (so Windows+Mac) now uses stdint.h
- exact set of macros involved in declaring Bool is reduced
- compiling with c90 mode really will break now (use gnu90 instead).