Fix the SSL/AsyncSocket "SSL_shutdown:shutdown while in init" problem
QE regression tests have caught a bug in the bora SSL/AsyncSocket
layer. A normal SSL connection could be abnormally terminated by some
other random connections.

Seems a common bug of the SSL usage according to the link below.
VMware is not the first one to hit it. :-)
https://marc.info/?l=openssl-users&w=4&r=1&s=shutdown+while+in+init&q=b

The cause is that the SSL error queue was not cleared before
calling the SSL IO functions like SSL_write,read,connect,accept.
Refer to the man page of SSL_get_error, Paragraph 2 of the Description.
Fixed the problem by adding addtional call to clear the error queue.

Found another problem that the system error is not preserved between
where it is set after the SSL IO call and the use of it to determine
whether to retry the SSL IO. This is probably due to tools has its
own logging which does not preserve the system error number.
Fixed the logging.