Hgfs Server: fix error result for new dir creation in the virtual folder
The Hgfs server code detects when a client tries to create a new folder in the
virtual folder containing the VMware Shares. This is not allowed as only the
host can add new Shares which appears in this folder. The current error code
returned to the client in these cases is access denied. This is only correct
when the new folder does not collide with an existing share. This stops clients
treating the virtual folder as a regular read/write physical folder. However,
for the case where the name collides, the access denied error is not correct,
as it does not tell the client the folder already exists. This causes some
applications to fail as they expect to see this error if the path component
exists as that is okay and expected. At this point the application will continue
to traverse the path and ensure the next component exists or if not create that.

To ensure that applications don't prematurely fail and can test that the share
still exists in the path the Hgfs server should return HGFS_ERROR_FILE_EXIST
and leave HGFS_ERROR_ACCESS_DENIED for all other cases.