Set X11 as the backend for gtk3 of open-vm-tools
Open-vm-tools defaults to gtk3.  For distros like Fedora 27, Wayland is
the default display server.  With no restriction on the backend, plugins
will try to use Wayland as the backend of Gtk+3.  As a result,
gdk_display_get_default() returns a Wayland display;
gdk_display_get_default_group() also returns a Wayland window.  Applying
GDK_WINDOW_XID() on the Wayland window will result in the crash reported
on recent Linux releases that default to a Wayland display server.

As X11-specific code is widely used in tools plugins, a migratation to
Wayland in the short term is not possible.  So, as a compromised solution,
plugins can be forced to run on XWayland, which is the compatible mode
of Wayland for legacy X11 clients.

gdk_set_allowed_backends() only applies when flag GTK3 is defined, and Gtk
version cwgreater than 3.10.