Set x11 as backend for gtk3 of open-vm-tools
Open-vm-tools defaults to gtk3, if available.  For distros where wayland is
the default display server, plugins will try to use wayland as the backend
of gtk3.  As a result, gdk_display_get_default() returns a wayland display,
and gdk_display_get_default_group() also returns a wayland window.  Then,
applying GDK_WINDOW_XID() on the wayland window will result in crashes.

As x11-specific code is widely used in tools plugins, as a compromized
solution, force plugins to run on XWayland, which is the compatible mode
of wayland for legacy X11 clients.

gdk_set_allowed_backends() only applies when flag GTK3 is defined,
and the gtk version is over 3.10.