Reduce Vigor transport socket draining overhead
A previous change in VigorPollImpl and asyncsocket to fix a memory
leak that causes a performance regression for parallel reconfiguration of
running VMs (the kind requiring FSR).  During an FSR, the connection
between hostd and the source vmx would drop and hostd would then
reconnect to the destination vmx.  One source of regression is that
the buffer has to be drained when a disconnect happens and the draining
logic may potentially unregister and re-register the recv callback
multiple times.  This change adds an asyncsocket function specifically
for draining recv that reduces the manipulation of callback.  It
reduces the regression by up to ~50% (the effect is more pronounced with
more VMs).