lib/file: File_GetSafeTmpDir is not aware of credentials changes
Some applications may masquerade their use (change their EUID).
Futhermore, each thread in an application can have a separate EUID.

Anytime a application or thread asked for its safe temporary directory,
it needs to get the same result regardless of how many times it asks.

Change File_GetSafeTmpDir to cache the EUID associated with the cached
values. If there is an EUID change, invalidate the existing cache
entries and recompute them. The recomputation process is stable, in
that it will obtain the same return any time it is called.

This way we get the benefit of the cache (performance), ensured
correctness (for applications that do not masquerade), and correctness
for those applications that do masquerade.