Ignore ENXIO errors with SyncDriver
A quiesced snapshot fails when the target VM has a bind mount.  The
problem is that the syncDriver gets an ENXIO error when it tries to
open the mount point.  To fix the problem, let the quiesce operation
skip a mount point when opening it results in an ENXIO error.

This change is based on pull request
https://github.com/vmware/open-vm-tools/pull/218 .