Skip specified file systems when doing a quiesced snapshot on Linux
Add a tools.conf setting "excludedFileSystems" that specifies one or
more file system mount points to be skipped over when performing a
quiesced snapshot on Linux guests.  The value of excludedFileSystems
is a comma-separated list of glob-style patterns as recognized by the
glib routines described here:

https://developer.gnome.org/glib/stable/glib-Glob-style-pattern-matching.html

With this change, when performing a quiesced snapshot, the sync driver
freeze routine removes from the list of mount points it is to process
any path that matches a pattern in the excludedFileSystems list.

In the course of testing the change, a bug was found in SyncDriverFreeze
in which it returned SD_ERROR rather than FALSE when the path list is
empty.

This change also includes some whitespace cleanup in syncDriverWin32.c.