Hgfs FUSE Client: enable the FUSE client for RHEL 7, Ubuntu 14.04 and others
The RHEL 7.x releases are all kernel 3.10 based. Currently the Hgfs FUSE
client will not be enabled for these earlier kernel versions, enabled only
for 4.0 kernels and later.  This means for default tools install on plaftorms
running those kernel versions the Hgfs kernel client must be installed and
used for Shared Folders.  The consequence of this means that the tar tools
installer must be run to provide the Hgfs kernel client on top of the OVT
install.

To remove the need to install the tar tools we must enable support for the
Hgfs FUSE client, which just requires the kernel version check to be modified
from 4.0.0 to 3.10.0.