vmcheck.c VmCheck_IsVirtualWorld(): Always check for a working backdoor.
The specific checks for Xen and VirtualPC hypervisors currently happen
only if the VMware hypervisor is not detected.  The test for a working
VMware backdoor is then done to avoid all other hypervisors.

In the case where running on a VMware hypervisor and the backdoor
channel has been disabled such as with

   monitor_control.restrict_backdoor = "TRUE"

vmtoolsd, vmware-toolbox-cmd and vmware-checkvm will crash when
attempting to get the version number of the installed VMware Tools.

Added an additional test to detect Linux KVM with the existing tests
for the Xen hypervisor and Microsoft Virtual PC.  Avoid checking for
a working backdoor if a non VMware hypervisor is seen.

Microsoft Hv checking to be added later and is tracked in a separate PR.
Oracle VirtualBox provides no unique CPUID vendor signature string.