[Wayland DnD] Part2: Open the uinput device file with root permission.
The device file /dev/uinput (or /dev/input/uinput) can only be accessed
by root account, so the dndcp plugin may failed to open this device file
if the current user is not root account.

A way to fix this issue is opening this device file with root account,
then pass the file descriptor to the sub task which is started with the
current account. An example for this solution is blockVM file system
device file.

This patch is part of the new feature 'Wayland support in Linux guest'.