[Wayland DnD] Part1: Add 'fakeMouse' module which is used to simulate the        mouse motion under Wayland
For Linux guest with X11, the VMTools uses the X system APIs to simulate
the mouse motion, but these X11 System APIs do not work in Wayland. Need
to pick up another method to simulate the mouse motion.

Another way to simulate the mouse motion is using uinput module.

uinput is a kernel module that makes it possible to emulate input devices
from userspace. By writing to /dev/uinput (or /dev/input/uinput) device,
 a process can create a virtual input device with specific capabilities.
Once this virtual device is created, the process can send events through
it, that will be delivered to userspace and in-kernel consumers.

Here is the link which contains more information about the uinput module:

https://www.kernel.org/doc/html/v4.12/input/uinput.html

This patch adds a new module named 'fakeMouse' which is used to simulate the mouse motion under Wayland.

This patch is part of the new feature 'Wayland support in Linux guest'.