Hgfs Server Plugin: only handle the shared folders client in hosted products
Only initialize the Shared Folders client redirector in hosted product
environments.

Details: Check if the tools system service started in a non-hosted product
and if so do not start the shared folders client redirector.
Check if the tools user service started in a non-hosted product product and if
so then skip reconnecting any Shared Folders mapped drives.