Fix locale issues for vgauth on Linux
setlocale() returns nothing unless we earlier force it
to be empty.

Change the default msgCatalog location on Linux so
the catalogs can be found without config changes in the normal
install area.

When doing catalog lookup, first chop off the charset
suffix.  Otherwise we first look for 'zh_CN.UTF-8', then
fall back to just 'zh' if that fails.  We support zh_TW and zh_CN,
but not a 'zh'.

Also fix things for NeoKylin, which appears to use LANG
but not setlocale().