Bump invalid PPN to 2^52 - 1
We are ready to bump INVALID_PPN value. Let us set it
to 2^52 - 1. This is more than enough to get over 4PB vRAM,
that requires 40-bit PPNs.

Another option was to set it to (uint64)-1, like the current value
of INVALID_PPN64, but it seems like there is no good reason
to do it. There are a few instances where INVALID_PPN is
treated as a regular PPN, e.g. we may convert it to PA and back.
2^52 - 1 will allow us to preserve that natural relationship
between invalid PPN and invalid PA.

All INVALID_PPN64 will have to be converted to INVALID_PPN
in follow up changes.