Change the balloon driver internals to 64-bits.
As a first step towards making the driver being able to balloon 16TB+,
let's change all the internals to use 64-bits values for the counters.
Initially, I wanted to only touch the windows driver, as this is the one
that we really care about, but since some code and datastructure are
shared, I had to modify pretty much all drivers.

The next step will be to change the drivers capabilities by adding
BALLOON_64_BIT_TARGET.